fixed the model issue
