+++ config/dbConfig.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-
+require("dotenv").config();
 mongoose.connect(process.env.MONGO_URL);
 
 const connection = mongoose.connection;
