+++ middlewares/authMiddleware.js
@@ -1,5 +1,7 @@
 const jwt = require("jsonwebtoken");
 
+  require("dotenv").config();
+
 module.exports = async (req, res, next) =>{
     try {
         const token = req.headers["authorization"].split(" ")[1];
