+++ models/userModel.js
@@ -4,7 +4,7 @@ const userSchema = new mongoose.Schema(
   {
     name: {
       type: String,
-      required: true,
+      required:true,
     },
     email: {
       type: String,
